new-author script added + refactored scripts structure
